requirements.txt
flask
tensorflow==2.15.0
pillow
numpy
gunicorn
gdown
